+++ CHANGELOG.md
@@ -8,7 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 
+- Bill Invoice Tax objects now support tax extensions.
 - MX Stamps for signatures from CFDI and SAT.
+- MX: extension for Place of Issue code: `mx-cfdi-issue-place` that replaces previous post code option in the supplier. Automatic normalization added.
 - `head` package now has `GetStamp` method to find a stamp by its provider from an array.
 - `num.Percentage` has `Base()` method to access base amount.